+++ packages/waggle/src/socket/constants.ts
@@ -19,5 +19,6 @@ export enum EventTypesServer {
   SUBSCRIBE_FOR_DIRECT_MESSAGE_THREAD = 'subscribeForDirectMessageThread',
   DIRECT_MESSAGE = 'directMessage',
   SUBSCRIBE_FOR_ALL_CONVERSATIONS = 'subscribeForAllConversations',
+  ASK_FOR_MESSAGES = 'askForMessages',
   SAVE_CERTIFICATE = 'saveCertificate'
 }