+++ packages/waggle/src/socket/events/message.ts
@@ -1,9 +1,0 @@
-import { EventTypesServer } from '../constants'
-
-export const message = (socket, message) => {
-  socket.emit(EventTypesServer.MESSAGE, message)
-}
-
-export const directMessage = (socket, message) => {
-  socket.emit(EventTypesServer.DIRECT_MESSAGE, message)
-}