+++ packages/waggle/src/socket/listeners/connection.ts
@@ -53,6 +53,12 @@ export const connections = (io, connectionsManager: ConnectionsManager) => {
     socket.on(EventTypesServer.SUBSCRIBE_FOR_ALL_CONVERSATIONS, async (conversations: string[]) => {
       await connectionsManager.subscribeForAllConversations(conversations)
     })
+    socket.on(
+      EventTypesServer.ASK_FOR_MESSAGES,
+      async ({ channelAddress, ids }: { channelAddress: string, ids: string[] }) => {
+        await connectionsManager.askForMessages(channelAddress, ids)
+      }
+    )
     socket.on(EventTypesServer.SAVE_CERTIFICATE, async (certificate: string) => {
       console.log('Received saveCertificate websocket event, processing.')
       await connectionsManager.saveCertificate(certificate)